Manual changes to fix lint
