+++ packages/@unimodules/react-native-adapter/src/index.ts
@@ -3,9 +3,7 @@ import NativeModulesProxy, { ProxyNativeModule } from './NativeModulesProxy';
 import { requireNativeViewManager } from './NativeViewManagerAdapter';
 import Platform from './Platform';
 import SyntheticPlatformEmitter from './SyntheticPlatformEmitter';
-
 // RCTDeviceEventEmitter pending https://github.com/necolas/react-native-web/pull/1402
-
 import { CodedError } from './errors/CodedError';
 import { UnavailabilityError } from './errors/UnavailabilityError';
 import { RCTDeviceEventEmitter } from './nativeEmitters';
