+++ packages/expo-asset/src/ts-declarations/react-native-asset-registry.d.ts
@@ -1,0 +1,5 @@
+declare module 'react-native/Libraries/Image/AssetRegistry' {
+  export type PackagerAsset = any;
+  export function registerAsset(asset: PackagerAsset): number;
+  export function getAssetByID(assetID: number): PackagerAsset | undefined;
+}
