+++ packages/expo-asset/src/ts-declarations/react-native-resolve-asset-source.d.ts
@@ -1,0 +1,11 @@
+declare module 'react-native/Libraries/Image/resolveAssetSource' {
+  import AssetSourceResolver, {
+    ResolvedAssetSource,
+  } from 'react-native/Libraries/Image/AssetSourceResolver';
+
+  export default function resolveAssetSource(source: any): ResolvedAssetSource;
+
+  export function setCustomSourceTransformer(
+    transformer: (resolver: AssetSourceResolver) => ResolvedAssetSource
+  ): void;
+}
