+++ packages/jest-expo-enzyme/package.json
@@ -54,6 +54,7 @@
   "devDependencies": {
     "@types/enzyme": "^3.10.3",
     "@types/enzyme-adapter-react-16": "^1.0.5",
+    "@types/jsdom": "^12.2.4",
     "babel-preset-expo": "~8.0.0",
     "expo-module-scripts": "^1.1.1",
     "react": "16.9.0"
