+++ packages/jest-expo/src/preset/setup.js
@@ -7,6 +7,8 @@
 const mockNativeModules = require('react-native/Libraries/BatchedBridge/NativeModules');
 
 const createMockConstants = require('./createMockConstants');
+const publicExpoModules = require('./expoModules');
+const internalExpoModules = require('./internalExpoModules');
 
 // window isn't defined as of react-native 0.45+ it seems
 if (typeof window !== 'object') {
@@ -31,8 +33,6 @@ Object.defineProperty(mockNativeModules, 'LinkingManager', {
   get: () => mockNativeModules.Linking,
 });
 
-const publicExpoModules = require('./expoModules');
-const internalExpoModules = require('./internalExpoModules');
 const expoModules = {
   ...publicExpoModules,
   ...internalExpoModules,
